Add storage/type_traits.h
