+++ be/src/column/datum_convert.cpp
@@ -4,6 +4,7 @@
 #include "gutil/strings/substitute.h"
 #include "runtime/mem_pool.h"
 #include "storage/olap_type_infra.h"
+#include "storage/type_traits.h"
 
 namespace starrocks::vectorized {
