+++ be/src/exprs/expr.h
@@ -28,6 +28,7 @@
 #include "column/chunk.h"
 #include "column/vectorized_fwd.h"
 #include "common/status.h"
+#include "common/statusor.h"
 #include "exprs/expr_context.h"
 #include "gen_cpp/Opcodes_types.h"
 #include "runtime/datetime_value.h"
