+++ be/src/exprs/vectorized/hyperloglog_functions.cpp
@@ -6,6 +6,7 @@
 #include "column/column_viewer.h"
 #include "column/object_column.h"
 #include "exprs/vectorized/unary_function.h"
+#include "types/hll.h"
 #include "util/phmap/phmap.h"
 
 namespace starrocks::vectorized {
