+++ be/src/runtime/data_stream_recvr.cc
@@ -33,6 +33,7 @@
 #include "gen_cpp/data.pb.h"
 #include "runtime/current_thread.h"
 #include "runtime/data_stream_mgr.h"
+#include "runtime/exec_env.h"
 #include "runtime/sorted_chunks_merger.h"
 #include "serde/protobuf_serde.h"
 #include "util/block_compression.h"
