+++ be/src/serde/column_array_serde.cpp
@@ -15,6 +15,7 @@
 #include "column/object_column.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/descriptors.h"
+#include "types/hll.h"
 #include "util/coding.h"
 #include "util/json.h"
 #include "util/phmap/phmap.h"
