+++ be/src/storage/column_vector.cpp
@@ -24,6 +24,7 @@
 #include <memory>
 
 #include "common/statusor.h"
+#include "storage/array_type_info.h"
 #include "storage/field.h"
 #include "storage/olap_common.h"
 #include "storage/olap_type_infra.h"
@@ -86,10 +87,9 @@ ScalarColumnVectorBatch<ScalarType>::ScalarColumnVectorBatch(const TypeInfoPtr& 
 ArrayColumnVectorBatch::ArrayColumnVectorBatch(const TypeInfoPtr& type_info, bool is_nullable, size_t init_capacity,
                                                Field* field)
         : ColumnVectorBatch(type_info, is_nullable), _item_offsets(1) {
-    auto array_type_info = down_cast<const ArrayTypeInfo*>(type_info.get());
     _item_offsets[0] = 0;
-    ColumnVectorBatch::create(init_capacity, field->get_sub_field(0)->is_nullable(), array_type_info->item_type_info(),
-                              field->get_sub_field(0), &_elements);
+    ColumnVectorBatch::create(init_capacity, field->get_sub_field(0)->is_nullable(),
+                              get_item_type_info(type_info.get()), field->get_sub_field(0), &_elements);
 }
 
 ArrayColumnVectorBatch::~ArrayColumnVectorBatch() = default;
