+++ be/src/storage/rowset/bitshuffle_page.h
@@ -37,6 +37,7 @@
 #include "storage/rowset/options.h"
 #include "storage/rowset/page_builder.h"
 #include "storage/rowset/page_decoder.h"
+#include "storage/type_traits.h"
 #include "storage/types.h"
 #include "types/date_value.hpp"
 #include "util/coding.h"
