+++ be/src/storage/rowset/frame_of_reference_page.h
@@ -25,6 +25,7 @@
 #include "storage/rowset/options.h"      // for PageBuilderOptions/PageDecoderOptions
 #include "storage/rowset/page_builder.h" // for PageBuilder
 #include "storage/rowset/page_decoder.h" // for PageDecoder
+#include "storage/type_traits.h"
 #include "util/frame_of_reference_coding.h"
 
 namespace starrocks {
