+++ be/src/storage/rowset/plain_page.h
@@ -26,6 +26,7 @@
 #include "storage/rowset/options.h"
 #include "storage/rowset/page_builder.h"
 #include "storage/rowset/page_decoder.h"
+#include "storage/type_traits.h"
 #include "storage/types.h"
 #include "util/coding.h"
 #include "util/faststring.h"
