+++ be/src/util/arrow/starrocks_column_to_arrow.cpp
@@ -5,6 +5,7 @@
 #include "column/type_traits.h"
 #include "common/statusor.h"
 #include "exec/vectorized/arrow_type_traits.h"
+#include "types/hll.h"
 #include "util/raw_container.h"
 
 namespace starrocks::vectorized {
